moved swagger.js into config folder
